Refactoring tests for osa_tool
