+++ tests/unit/osa_tool/conftest.py
@@ -1,0 +1,6 @@
+import pytest
+
+
+@pytest.fixture
+def repo_url():
+    return "https://github.com/username/repo-name"
